tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/protocol/005.xhtml

<?xml version="1.0" encoding="utf-8"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Invalid characters in protocol scheme and isProtocolHandlerRegistered</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script type="application/ecmascript">
test(function() {
  var scheme = 'web+CústomPrótocolFíve', url = location.href.replace(/\/[^\/]*$/, "") + '/%s';
  navigator.registerProtocolHandler(scheme, url, 'Ignore dialog or decline it');
  assert_equals(navigator.isProtocolHandlerRegistered(scheme, url), 'new');
});
</script>
</head>
<body>
<div id="log"></div>
</body>
</html>
